修改项目结构，修改模块和包名
