--- WF-Consumer/pom.xml
@@ -1,33 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>org.radrso.workflow</groupId>
-        <artifactId>Workflow-Engine</artifactId>
-        <version>1.0-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>WF-Consumer</artifactId>
-    <packaging>war</packaging>
-    <name>Workflow-Engine Consumer Maven Webapp</name>
-
-    <dependencies>
-
-        <dependency>
-            <groupId>org.radrso.workflow</groupId>
-            <artifactId>WF-Plugins</artifactId>
-            <version>1.0-SNAPSHOT</version>
-        </dependency>
-
-        <dependency>
-            <groupId>org.radrso.workflow</groupId>
-            <artifactId>WF-Interfaces</artifactId>
-            <version>1.0-SNAPSHOT</version>
-        </dependency>
-
-    </dependencies>
-
-    <build>
-        <finalName>Workflow-Engine-Consumer</finalName>
-    </build>
-</project>
