--- WF-Interfaces/src/main/java/org/radrso/entities/config/items/Judge.java
@@ -1,28 +1,0 @@
-package org.radrso.entities.config.items;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-
-import java.io.Serializable;
-
-/**
- * Created by raomengnan on 16-12-8.
- * compare  要比较的对象
- * judge    判断条件  >  =  <  >=  <=
- * next     if true, next
- */
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@ToString
-public class Judge implements Serializable{
-    private String compute;
-    private String computeWith;
-    private String expression;
-    private String next;
-    private Transfer passTransfer;
-    private Transfer nopassTransfer;
-}
