--- WF-Plugins/src/main/java/org/radrso/plugins/requests/entity/Method.java
@@ -1,20 +1,0 @@
-package org.radrso.plugins.requests.entity;
-
-import java.io.Serializable;
-
-/**
- * Created by raomengnan on 16-12-9.
- */
-public enum Method implements Serializable{
-    GET("Get"),
-    POST("Post"),
-    PUT("Put"),
-    DELETE("Delete");
-
-    String methodName;
-    Method(String name) {
-        this.methodName = name;
-    }
-
-    public String value(){return this.methodName;}
-}
