--- WF-Plugins/src/main/java/org/radrso/plugins/requests/entity/exceptions/impl/RequestException.java
@@ -1,30 +1,0 @@
-package org.radrso.plugins.requests.entity.exceptions.impl;
-
-import org.radrso.plugins.requests.entity.exceptions.ResponseCode;
-import org.radrso.plugins.requests.entity.exceptions.WFException;
-
-/**
- * Created by raomengnan on 16-12-9.
- */
-
-public class RequestException extends WFException {
-    public RequestException(){
-        super();
-    }
-
-    public RequestException(ResponseCode code){
-        super( code);
-    }
-
-    public RequestException(String msg, ResponseCode code){
-        super(msg, code);
-    }
-
-    public RequestException(String msg, ResponseCode code, Throwable cause){
-        super(msg, code, cause);
-    }
-
-    public RequestException(ResponseCode code, Throwable cause) {
-        super(code, cause);
-    }
-}
