+++ WF-Public/src/main/java/org/radrso/workflow/entities/config/WorkflowConfig.java
@@ -1,0 +1,28 @@
+package org.radrso.workflow.entities.config;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+import org.radrso.workflow.entities.config.items.Step;
+
+import java.io.Serializable;
+import java.util.Date;
+import java.util.List;
+
+/**
+ * Created by raomengnan on 16-12-8.
+ */
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@ToString
+public class WorkflowConfig implements Serializable{
+    private String application;
+    private String id;
+    private String header;
+    private Date startTime;
+    private Date stopTime;
+    private List<Step> steps;
+}
