+++ WF-Public/src/main/java/org/radrso/workflow/entities/config/items/Transfer.java
@@ -1,0 +1,27 @@
+package org.radrso.workflow.entities.config.items;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+import java.io.Serializable;
+import java.util.List;
+
+/**
+ * Created by raomengnan on 16-12-8.
+ * 状态转移函数
+ * wait 表示等待的时间
+ */
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@ToString
+public class Transfer implements Serializable{
+    private List<InputItem> input;
+    private Judge judge;
+    private int wait; //minutes
+    private String to;
+    private List<String> scatters;
+}
