+++ WF-Public/src/main/java/org/radrso/workflow/entities/exceptions/ConfigReadException.java
@@ -1,0 +1,12 @@
+package org.radrso.workflow.entities.exceptions;
+
+/**
+ * Created by raomengnan on 17-1-14.
+ */
+public class ConfigReadException extends Exception{
+    public static final String PARAM_RESOLVER_EXCEPTION = "param resolver exception";
+    private String msg;
+    public ConfigReadException(String msg){
+        super(PARAM_RESOLVER_EXCEPTION + String.format("[%s]", msg));
+    }
+}
