+++ pom.xml
@@ -6,9 +6,9 @@
     <packaging>pom</packaging>
     <version>1.0-SNAPSHOT</version>
     <modules>
-        <module>WF-Interfaces</module>
+        <module>WF-Public</module>
         <module>WF-Provider</module>
-        <module>WF-Consumer</module>
+        <module>WF-Service</module>
         <module>WF-Plugins</module>
     </modules>
     <name>Workflow-Engine Maven Webapp</name>
